time-controller.py
prepare a simple code ready to evaluate the end time of a period of time, given as a number of minutes. The start time is a given as a pair of hours (0...23) and minutes (0...59). The result has to be printed on the console. for example, if an event starts at 12:17 and lasts 59 minutes, it will end at 13:16.